requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
django-environ==0.8.1
django-filter==21.1
django-widget-tweaks==1.4.12
gunicorn==20.1.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
whitenoise==6.1.0
